Extracted 'sleep', 'enter', and 'wake' commands to separate files.
